Moved the omniauth stub method from the feature test files to the fixtures folder.
